Refactor Login page
