--- src/modules/login/actions/login-actions.constants.js
@@ -1,2 +1,0 @@
-export const LOGIN = "LOGIN";
-export const LOGIN_TWO_FACTOR = "LOGIN_TWO_FACTOR";
