+++ src/modules/login/components/login/login-form.js
@@ -11,27 +11,23 @@ import validationSchema from "./login-form.validators";
 const LoginForm = ({ t, isSubmitting, handleSubmit, error }) => {
   return (
     <Form
-      className={"login-form"}
+      className="login-form"
       id="loginForm"
       onSubmit={handleSubmit}
       noValidate
     >
       <GVFormikField
-        id={"loginEmail"}
         type="email"
         name="email"
         placeholder={t("login-form.placeholder.email")}
         autoComplete="email"
-        className={"login-input gv-text-field"}
         component={GVTextField}
       />
       <GVFormikField
-        id={"loginPassword"}
         type="password"
         name="password"
         placeholder={t("login-form.placeholder.password")}
         autoComplete="current-password"
-        className={"login-input gv-text-field"}
         component={GVTextField}
       />
       <a href={FORGOT_PASSWORD_ROUTE}>{t("login-form.forgot")}</a>
