+++ src/modules/login/reducers/login-reducers.js
@@ -1,9 +1,8 @@
 import { combineReducers } from "redux";
 
 import apiReducerFactory from "../../../shared/reducers/api-reducer/api-reducer";
+import { LOGIN } from "../actions/login-actions";
 import twoFactorReducer from "./two-factor-reducer";
-
-import { LOGIN } from "../actions/login-actions.constants";
 
 const loginReducer = combineReducers({
   login: apiReducerFactory({ apiType: LOGIN }),
