+++ src/modules/login/service/login-service.js
@@ -5,8 +5,8 @@ import authActions from "../../../actions/auth-actions";
 import authService from "../../../services/auth-service";
 import clearDataActionFactory from "../../../shared/actions/clear-data.factory";
 import history from "../../../utils/history";
+import { LOGIN, LOGIN_TWO_FACTOR } from "../actions/login-actions";
 import loginActions from "../actions/login-actions";
-import { LOGIN, LOGIN_TWO_FACTOR } from "../actions/login-actions.constants";
 import { RECOVERY_CODE, TWO_FACTOR_CODE } from "../login.constants";
 
 const login = (loginData, from, onCatch) => dispatch => {
