+++ src/pages/login/login.scss
@@ -91,19 +91,8 @@
     }
   }
 
-  .login-input {
-    margin-bottom: 16px;
-  }
-
   .login-button {
     margin-top: 50px;
-  }
-
-  .gv-text-field__wrapper {
-    width: 100%;
-    margin-left: 0;
-    margin-right: 0;
-    margin-top: 0;
   }
 
   .gv-text-field__error {
