+++ src/pages/login/two-factor/two-factor.js
@@ -1,4 +1,4 @@
-import TwoFactorCodeContainer from "modules/login/components/two-factor-code-container/two-factor-code-container";
+import TwoFactorCodeContainer from "modules/login/components/two-factor-code/two-factor-code-container";
 import { LOGIN_ROUTE_TWO_FACTOR_RECOVERY_ROUTE } from "pages/login/login.routes";
 import React from "react";
 import { Link } from "react-router-dom";
